separate sym
