+++ src/lib.rs
@@ -11,10 +11,10 @@ mod dynamic;
 mod enum_based;
 pub mod float_ops;
 pub mod ops;
-pub mod variables;
+pub mod symbols;
 
 pub use float_ops::{ExNumConsts, ExNumOps};
-pub use variables::*;
+pub use symbols::variables::*;
 
 /// Trait for Symbol using dynamic.
 pub trait DynamicSymbol<Out, In: ?Sized>: Any + Display {
@@ -169,181 +169,5 @@ where
 
     fn as_any(&self) -> &(dyn Any) {
         self
-    }
-}
-
-/// Marker for Unary Operation used in [`UnarySym`](`crate::UnarySym`).
-pub trait UnaryOp {
-    ///Returns the op name.
-    fn op_name<'a>() -> &'a str {
-        let s = std::any::type_name::<Self>();
-        debug_assert!(s.ends_with("Op"));
-        let op_name = &s[..s.len() - 2];
-        op_name
-    }
-
-    ///Formats the expression to display.
-    fn format_expression(
-        f: &mut fmt::Formatter<'_>,
-        inner: impl FnOnce(&mut fmt::Formatter<'_>) -> Result<(), fmt::Error>,
-    ) -> Result<(), fmt::Error> {
-        f.write_fmt(format_args!("{}( ", Self::op_name()))?;
-        inner(f)?;
-        f.write_str(")")
-    }
-}
-
-/// [`Symbol`](`crate::Symbol`) represent Unary Operation.
-#[derive(Debug, PartialEq, Eq)]
-pub struct UnarySym<Op, Sym, Out, In: ?Sized = Out> {
-    op: Op,
-    sym: Sym,
-    po: PhantomData<Out>,
-    pi: PhantomData<In>,
-}
-
-impl<Op, Sym, Out, In: ?Sized> UnarySym<Op, Sym, Out, In>
-where
-    Op: UnaryOp,
-    Sym: Symbol<Out, In>,
-{
-    fn new(sym: Sym) -> Self
-    where
-        Op: Default,
-    {
-        UnarySym {
-            op: Op::default(),
-            sym: sym,
-            po: PhantomData,
-            pi: PhantomData,
-        }
-    }
-
-    fn new_with_op(op: Op, sym: Sym) -> Self {
-        UnarySym {
-            op: op,
-            sym: sym,
-            po: PhantomData,
-            pi: PhantomData,
-        }
-    }
-}
-
-impl<Op, Sym, Out, In: ?Sized> Clone for UnarySym<Op, Sym, Out, In>
-where
-    Op: UnaryOp + Clone,
-    Sym: Symbol<Out, In>,
-{
-    fn clone(&self) -> Self {
-        UnarySym {
-            op: self.op.clone(),
-            sym: self.sym.clone(),
-            po: PhantomData,
-            pi: PhantomData,
-        }
-    }
-}
-
-impl<Op, Sym, Out, In: ?Sized> From<Sym> for UnarySym<Op, Sym, Out, In>
-where
-    Op: UnaryOp + Default,
-    Sym: Symbol<Out, In>,
-{
-    #[inline]
-    fn from(v: Sym) -> Self {
-        UnarySym {
-            op: Op::default(),
-            sym: v,
-            po: PhantomData,
-            pi: PhantomData,
-        }
-    }
-}
-
-/// Marker for Binary Operation used in [`BinarySym`](`crate::BinarySym`).
-pub trait BinaryOp {
-    /// Symbol for this expression.
-    fn op_symbol() -> Option<&'static str> {
-        None
-    }
-
-    fn format_expression(
-        f: &mut fmt::Formatter<'_>,
-        left: impl FnOnce(&mut fmt::Formatter<'_>) -> Result<(), fmt::Error>,
-        right: impl FnOnce(&mut fmt::Formatter<'_>) -> Result<(), fmt::Error>,
-    ) -> Result<(), fmt::Error> {
-        let s = std::any::type_name::<Self>();
-        debug_assert!(s.ends_with("Op"));
-        let op_name = &s[..s.len() - 2];
-        if let Some(sym) = Self::op_symbol() {
-            left(f)?;
-            f.write_fmt(format_args!(" {} ", sym))?;
-            right(f)
-        } else {
-            f.write_fmt(format_args!("{}( ", op_name))?;
-            left(f)?;
-            right(f)?;
-            f.write_str(")")
-        }
-    }
-}
-
-/// [`Symbol`](`crate::Symbol`) represent Binary Operation.
-#[derive(Debug, PartialEq, Eq)]
-pub struct BinarySym<Op, Sym1, Sym2, Out, In: ?Sized = Out> {
-    op: Op,
-    sym1: Sym1,
-    sym2: Sym2,
-    po: PhantomData<Out>,
-    pi: PhantomData<In>,
-}
-
-impl<Op: BinaryOp, Sym1: Symbol<Out, In>, Sym2: Symbol<Out, In>, Out, In: ?Sized>
-    BinarySym<Op, Sym1, Sym2, Out, In>
-{
-    pub fn new_with_op(op: Op, sym1: Sym1, sym2: Sym2) -> Self {
-        BinarySym {
-            op: op,
-            sym1: sym1,
-            sym2: sym2,
-            po: PhantomData,
-            pi: PhantomData,
-        }
-    }
-
-    pub fn new(sym1: Sym1, sym2: Sym2) -> Self
-    where
-        Op: Default,
-    {
-        BinarySym {
-            op: Op::default(),
-            sym1: sym1,
-            sym2: sym2,
-            po: PhantomData,
-            pi: PhantomData,
-        }
-    }
-}
-
-impl<Op, Sym1, Sym2, Out, In: ?Sized> Clone for BinarySym<Op, Sym1, Sym2, Out, In>
-where
-    Op: BinaryOp + Clone,
-    Sym1: Symbol<Out, In> + Clone,
-    Sym2: Symbol<Out, In> + Clone,
-{
-    fn clone(&self) -> Self {
-        Self::new_with_op(self.op.clone(), self.sym1.clone(), self.sym2.clone())
-    }
-}
-
-impl<Op, Sym1, Sym2, Out, In: ?Sized> From<(Sym1, Sym2)> for BinarySym<Op, Sym1, Sym2, Out, In>
-where
-    Op: BinaryOp + Default,
-    Sym1: Symbol<Out, In>,
-    Sym2: Symbol<Out, In>,
-{
-    #[inline]
-    fn from(v: (Sym1, Sym2)) -> Self {
-        BinarySym::new(v.0, v.1)
     }
 }
