+++ src/symbols/variables/d_monomial.rs
@@ -1,3 +1,4 @@
+use crate::symbols::*;
 use crate::*;
 use core::{
     any::Any,
