+++ src/symbols/variables/d_variable.rs
